Core: only store persistent changes if there are changes

+++ Utils.py
@@ -322,9 +322,11 @@ def get_options() -> Settings:
     return get_settings()
 
 
-def persistent_store(category: str, key: str, value: typing.Any):
+def persistent_store(category: str, key: str, value: typing.Any, force_store: bool = False):
     path = user_path("_persistent_storage.yaml")
     storage = persistent_load()
+    if not force_store and category in storage and key in storage[category] and storage[category][key] == value:
+        return  # no changes necessary
     category_dict = storage.setdefault(category, {})
     category_dict[key] = value
     with open(path, "wt") as f:
